+++ CHANGELOG.md
@@ -6,7 +6,13 @@ All notable changes to this project will be documented in this file. This projec
 
 [Full Changelog](https://github.com/Shatur95/optimus-manager-qt/compare/1.2.0...HEAD)
 
-_No unreleased changes yet_
+**Added**
+
+-   Finnish translation.
+
+**Changed**
+
+-   Update Spanish translation.
 
 ## [1.2.0](https://github.com/Shatur95/optimus-manager-qt/tree/1.2.0) (2019-05-12)